actual impl?
bench: 2157968

bench: 2157968

+++ src/engine/search/search.cpp
@@ -476,9 +476,9 @@ int32_t Worker::search(
 
   if (!stack[ply].excluded) {
     ttable.add(
-      {.hash = (uint16_t)(position.full_hash() >> 48),
+      {.hash = position.full_hash() >> 48,
        .move = PV.moves[0],
-       .score = static_cast<int16_t>(best),
+       .score = best,
        .depth = static_cast<uint8_t>(depth),
        .node_type = node_type,
        .tt_PV = is_PV || tt_entry.tt_PV}
@@ -597,9 +597,9 @@ int32_t Worker::quiesce(Position& position, Line& PV, int16_t ply, int32_t alpha
   }
 
   ttable.add(
-    {.hash = static_cast<uint16_t>(position.full_hash()),
+    {.hash = position.full_hash(),
      .move = PV.moves[0],
-     .score = static_cast<int16_t>(best),
+     .score = best,
      .depth = 0,
      .node_type = node_type,
      .tt_PV = is_PV || tt_entry.tt_PV}
